wasm: Add emscripten/WebAssembly as an OS
Add code to os.h and platform.h to detect when emscripten is being used
to compile libflashrom. The architecture will be 'wasm32' and the OS
will be 'emscripten'. This will be used to limit the programmers
supported to be ones that can work on the web, i.e. webUSB ones like
raiden_debug_spi.

Change-Id: I9b78e2bb46882311b534005b0042c92d45c11f35
Signed-off-by: Stephen Boyd <swboyd@chromium.org>